Promijenjena lokacija fake objekta
